requirements.txt
Flask==2.2.5
python-dotenv==1.0.1
requests==2.32.3
numpy<2,>=1.18
h3<4,>=3.7.6
timezonefinder==6.2.0
timezonefinder==6.2.0
pytz==2024.1
flask-cors==4.0.1
apify-client==1.6.1